+++ src/test/java/org/jsoup/parser/XmlTreeBuilderTest.java
@@ -286,6 +286,20 @@ public class XmlTreeBuilderTest {
         assertEquals("<a bB1-_:.=\"foo\" _9_=\"bar\" xmlns:p1=\"qux\">One</a>", out); // first is same, second coerced
     }
 
+    @Test void normalizesAttributeNamesWithInvalidXmlStarts() {
+        // digits and hyphens are valid within XML names, but not at the start
+        Document doc = Jsoup.parse("<root 1a='one' -a='two' />", Parser.xmlParser());
+
+        assertEquals("<root _1a=\"one\" _-a=\"two\" />", doc.html());
+    }
+
+    @Test void makesRepairedXmlAttributeNameUnique() {
+        // valid names are preserved and repaired collisions receive underscore prefixes until unique
+        Document doc = Jsoup.parse("<root 1a='one' _1a='two' __1a='three' />", Parser.xmlParser());
+
+        assertEquals("<root ___1a=\"one\" _1a=\"two\" __1a=\"three\" />", doc.html());
+    }
+
     @Test void customTagsAreFlyweights() {
         String xml = "<foo>Foo</foo><foo>Foo</foo><FOO>FOO</FOO><FOO>FOO</FOO>";
         Document doc = Jsoup.parse(xml, Parser.xmlParser());
@@ -713,8 +727,17 @@ public class XmlTreeBuilderTest {
         // we infer that empty els can be represented with self-closing if seen in parse
     }
 
-    @Test public void xmlParserHasUnlimitedDepthByDefault() {
+    @Test public void xmlParserHasDefaultDepth() {
         Parser parser = Parser.xmlParser();
+        Document doc = Jsoup.parse(deepXml(600), "", parser);
+        Element target = doc.selectFirst("target");
+        assertNotNull(target);
+        assertEquals(512, parser.getMaxDepth());
+        assertEquals(parser.getMaxDepth(), depth(target));
+    }
+
+    @Test public void xmlParserCanUseUnlimitedDepth() {
+        Parser parser = Parser.xmlParser().setMaxDepth(Integer.MAX_VALUE);
         Document doc = Jsoup.parse(deepXml(600), "", parser);
         Element target = doc.selectFirst("target");
         assertNotNull(target);
@@ -727,6 +750,18 @@ public class XmlTreeBuilderTest {
         Element target = doc.selectFirst("target");
         assertNotNull(target);
         assertEquals(parser.getMaxDepth(), depth(target));
+    }
+
+    @Test void deepEndTagClosesElement() {
+        StringBuilder xml = new StringBuilder("<root><ancestor>");
+        for (int i = 0; i < 300; i++) {
+            xml.append("<n>");
+        }
+        xml.append("</ancestor><target />");
+
+        Document doc = Jsoup.parse(xml.toString(), "", Parser.xmlParser());
+        Element target = doc.expectFirst("target");
+        assertEquals("root", target.parent().normalName());
     }
 
     @Test void namespaceScopeTracksDepthPruning() {